added reset password wit temporary password feature

+++ src/main/java/com/techriff/userdetails/service/UsersDetailsService.java
@@ -18,11 +18,12 @@ import org.json.simple.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Page;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 
+import com.techriff.mail.service.EmailService;
 import com.techriff.userdetails.Exception.AgeLimitNotReachedException;
 import com.techriff.userdetails.Exception.DuplicateResourceException;
 import com.techriff.userdetails.Exception.ResourceNotFoundException;
@@ -39,7 +40,6 @@ import com.techriff.userdetails.entity.UserRoleMap;
 import com.techriff.userdetails.entity.UserRoleMapPK;
 import com.techriff.userdetails.entity.Users;
 import com.techriff.userdetails.entity.UsersRole;
-import com.techriff.userdetails.mail.service.EmailService;
 import com.techriff.userdetails.pages.UsersPage;
 import com.techriff.userdetails.pages.UsersSearchCriteria;
 import com.techriff.userdetails.repository.AddressRepository;
@@ -71,12 +71,14 @@ public class UsersDetailsService {
     private AddressRepository addressRepository;
     @Autowired
     private AddressTypeRepository addressTypeRepository;
+    @Autowired
+    private PasswordEncoder passwordEncoder;
     
     @Autowired private AddressService addressService;
     @Autowired PasswordResetTokenRepository passwordResetTokenRepository;
     @Autowired TemporaryPasswordRepository  tempRepo;
     
-   
+
 
     private EmailService emailService;
     //private static Logger logger = LogManager.getLogger(UsersDetailsService.class);
@@ -239,9 +241,13 @@ public class UsersDetailsService {
         return age;
     }
 
-    public String deleteuserById(int id) {
+    public String deleteuserById(int id) throws ResourceNotFoundException {
+    	if (usersRepository.findById(id)!=null)
+    	{
         usersRepository.deleteById(id);
         return "User Deleted with id: " + id;
+    	}
+    	else throw new ResourceNotFoundException("Users not present"+id);
     }
 
     @SuppressWarnings("unchecked")
@@ -329,20 +335,16 @@ public class UsersDetailsService {
     }
     //save data without file
     private Users savingUserDetail(Users users) throws Exception {
-       
+       // Users requestUser=users;
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
         Calendar dob = Calendar.getInstance();
         String dobOfUser = sdf.format(users.getDob());
         dob.setTime(sdf.parse(dobOfUser));
         int age = getAge(dob);
         if (age >= 18) {
-            users.setPassword(
-                bCryptPasswordEncoder.encode(users.getPassword())
-                
-            );
 
         
-            
+            users.setPassword(passwordEncoder.encode(users.getPassword()));
             List<Address> addresses = users.getAddress();
             List<AddressType> exitingAddressType=addressTypeRepository.findAll();
             for (Address data : addresses)
@@ -409,8 +411,12 @@ public class UsersDetailsService {
 
     public void createTemporaryPasswordForUser(Users existingUsers, String password) {
         
-        TemporaryPassword tempPassword=new TemporaryPassword(password,existingUsers);
+        TemporaryPassword tempPassword=new TemporaryPassword(passwordEncoder.encode(password) ,existingUsers);
+      //  TemporaryPassword tempPassword=new TemporaryPassword(password ,existingUsers);
+        tempPassword.setFlag(true);
         tempRepo.save(tempPassword);
+        existingUsers.setPasswordType("Temporary");
+        usersRepository.save(existingUsers);
         
     }
